Remove the last commit. Was wrong. Also this is not right.

+++ ExamApp/Gtin.cs
@@ -9,6 +9,7 @@ namespace ExamApp
     public class GTin
     {
         public string Code { get; set; }
-        public decimal Price { get; set; }
+        public int Price { get; set; }
+
     }
 }
